feat: BottomLineTextField Reset Button Event 추가 [IOS-26]
